enhance GPX track finder to need less config in the map
The whole thing is still pretty hacky, and the matching of map data
to video tracks is very brittle. This needs a full redesign that
takes the complicated nature of "track side arms" and only partial
video coverage into account.

But alas, for now we gain GPX with less config as long as the
nodes have target=X tags. This commit will break the build until
the map is updated due to the RSW targets not being setup properly.