increase buffer size when video streaming
Weirdly enough sometimes segments load slowly, sometimes fast. This
happens even when requesting the same segment repeatedly, so it
doesn't seem to be a caching or general "ISP is overloaded" issue.

This stop gap measure at least allows to pause the video to cache
up to 15s, maybe it hides the issue somewhat.