fix video and GPS indicator drifting apart
The reason was that the field printed by mediainfo was the video
duration, whereas ffmpeg seemingly worked on the format duration.
The small differences added up over time, which was especially
noticable on long videos.

Technically this is a breaking change, but it's not covered by the
renderer version. The video length which causes this is handled by
source and was assumed to be "static". Hence fixing this everywhere
would require all videos to change their hashes. Rerendering is not
required, but tracking old/new hash will be a hassle.

Instead, whenever re-rendering, manual updating of the videos with
an old, cached timestamp needs to be done:

```elixir
track = <some track here>
track.videos
|> Enum.map(&elem(&1,0))
|> Enum.map(fn x -> "videos/source/#{x}.gpx" end)
|> Enum.filter(fn x -> File.exists?("#{x}_backup_without_time") end)
|> Enum.each(fn x-> File.copy("#{x}_backup_without_time", x) end)
```