make video metadata cache dependent on video constants
When changing the constants for one-off videos, the metadata file
needs to be re-created, since it also contains computed values
like PTS correction.

The proper way would be to move these computed values out of the
cache and calculate them on the fly. However, the constants change
rarely, so the refactoring doesn't seem worth it.